chore: rearange pages folder structure
